requirements.txt
Click==7.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.18.0
six==1.13.0
SQLAlchemy==1.3.12
torch==1.3.1
torchvision==0.4.2
Werkzeug==0.16.0
Pillow>=7.1.0
